requirements.txt
Pillow
pi-heif
